Changes several functions to pure functions
