+++ FrontEnd/source/index.html
@@ -32,7 +32,7 @@
         <button id="reset-button" onclick="resetGame()" style="display: none;">Reset Game</button>
     </div>
     <div id="indicator-area" class="flex-container">
-        <div id="win-count-area">
+        <div id="win-count-area" style="display: none;">
             <p id="win-count-label" class="flex-container">Win Count:</p>
             <div id="win-counters" class="flex-container">
                 <div id="red-counter" class="counter">
