+++ package.json
@@ -10,14 +10,17 @@
         "type": "git",
         "url": "git+https://github.com/athavan9/Connect4.git"
     },
-    "author": "Athavan & Taylor",
+    "author": "Athavan",
     "license": "ISC",
     "bugs": {
         "url": "https://github.com/athavan9/Connect4/issues"
     },
     "homepage": "https://github.com/athavan9/Connect4#readme",
     "dependencies": {
-        "jest": "^26.4.2"
+        "cors": "^2.8.5",
+        "express": "^4.17.1",
+        "jest": "^26.4.2",
+        "nodemon": "^2.0.4"
     },
     "devDependencies": {
         "eslint": "^7.8.1",
